Add Dangerzone .repo file
The goal of having the .repo here is to allow users to install it with
one command:

  sudo yum-config-manager --add-repo=https://packages.freedom.press/yum-tools-prod/dangerzone/dangerzone.repo

The key is set to a file in the user's file system. This is because on
the instructions the users will have downloaded the key from a
keyserver. This way the installation security depends on both the
keyserver and the repo.

See [1] for the corresponding PR updating the instructions on the
Dangerzone application.

[1]: https://github.com/freedomofpress/dangerzone/pull/385/